backend/backup_files/requirements.txt
openai
faiss-cpu
# Добавьте сюда другие зависимости, если они используются в qa.py или buildIndex.py
